Fix Solr command tests
