+++ catalog/solr/catalog-solr-commands/src/test/java/org/codice/ddf/commands/solr/SolrCommandTest.java
@@ -33,6 +33,10 @@ import org.apache.solr.client.solrj.response.RequestStatusState;
 import org.apache.solr.cloud.MiniSolrCloudCluster;
 import org.apache.solr.common.SolrInputDocument;
 import org.apache.solr.common.util.NamedList;
+import org.junit.After;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.rules.TemporaryFolder;
@@ -67,6 +71,34 @@ public abstract class SolrCommandTest {
   protected static String protocols;
 
   protected static MiniSolrCloudCluster miniSolrCloud;
+
+  @BeforeClass
+  public static void beforeClass() throws Exception {
+    setDdfHome();
+    setDdfEtc();
+    createDefaultMiniSolrCloudCluster();
+    addDocument("1");
+  }
+
+  @AfterClass
+  public static void afterClass() throws Exception {
+    if (miniSolrCloud != null) {
+      miniSolrCloud.getSolrClient().close();
+      miniSolrCloud.shutdown();
+    }
+  }
+
+  @Before
+  public void setUp() throws Exception {
+    setupSolrClientType(SolrCommands.CLOUD_SOLR_CLIENT_TYPE);
+    consoleOutput = new ConsoleOutput();
+    consoleOutput.interceptSystemOut();
+  }
+
+  @After
+  public void tearDown() {
+    consoleOutput.resetSystemOut();
+  }
 
   protected static Path getBaseDirPath() {
     return baseDir.getRoot().toPath();
